Refactoring tasit-action test suites to use global hooks
