+++ packages/tasit-action/package.json
@@ -18,7 +18,7 @@
   },
   "scripts": {
     "prepare": "rm -rf dist/* && npx babel src --out-dir dist --ignore src/*.test.js,src/**/*.test.js",
-    "test": "npm run lint && npx mocha \"src/**/*.test.js\" --recursive --require @babel/register --require src/testHelpers/mochaSetup.js --no-timeouts",
+    "test": "npm run lint && npx mocha \"src/**/*.test.js\" --recursive --require @babel/register --file src/testHelpers/mochaSetup.js --no-timeouts",
     "lint": "npx prettier src/* --write"
   },
   "bugs": {
@@ -34,10 +34,8 @@
     "@babel/plugin-proposal-class-properties": "^7.2.3",
     "@babel/preset-env": "^7.1.6",
     "@babel/register": "^7.0.0",
-    "bn.js": "^4.11.8",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "chai-bn": "^0.1.1",
     "mocha": "^5.2.0",
     "sinon": "^7.2.2",
     "tasit-account": "^0.0.9"
