+++ packages/tasit-action/src/ethers.test.js
@@ -9,12 +9,8 @@ const sampleContractAddress = "0x6C4A015797DDDd87866451914eCe1e8b19261931";
 
 let wallet;
 let sampleContract;
-let testcaseSnaphotId;
 
 describe("ethers.js", () => {
-  const provider = new ethers.providers.JsonRpcProvider();
-  provider.pollingInterval = 50;
-
   beforeEach("instantiate provider, wallet and contract objects", async () => {
     const privateKey =
       "0x11d943d7649fbdeb146dc57bd9cfc80b086bfab2330c7b25651dbaf382392f60";
@@ -29,16 +25,6 @@ describe("ethers.js", () => {
       wallet
     );
     expect(sampleContract.address).to.equal(sampleContractAddress);
-
-    testcaseSnaphotId = await createSnapshot(provider);
-  });
-
-  afterEach("revert blockchain snapshot", async () => {
-    await revertFromSnapshot(provider, testcaseSnaphotId);
-
-    // Note: Without this the test suite is breaking.
-    // It is still unclear why
-    await mineBlocks(provider, 1);
   });
 
   it("should instatiate contract object using human-readable ABI", async () => {
